+++ package.json
@@ -17,13 +17,15 @@
     "@reduxjs/toolkit": "^1.9.5",
     "axios": "^1.5.0",
     "dotenv": "^16.3.1",
+    "formik": "^2.4.5",
     "modern-normalize": "^2.0.0",
     "notiflix": "^3.2.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.1.2",
     "react-router-dom": "^6.16.0",
-    "redux-persist": "^6.0.0"
+    "redux-persist": "^6.0.0",
+    "yup": "^1.2.0"
   },
   "devDependencies": {
     "@types/react": "^18.2.15",
